Completed up to the create the view model section on the sort, filter, page and group page

+++ ContosoUniversity/ContosoUniversity/Controllers/StudentsController.cs
@@ -20,9 +20,53 @@ namespace ContosoUniversity.Controllers
         }
 
         // GET: Students
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(
+			string sortOrder, 
+			string searchString,
+			string currentFilter,
+			int? page)
         {
-            return View(await _context.Students.ToListAsync());
+			ViewData["CurrentSort"] = sortOrder;
+			ViewData["NameSortParm"] = string.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
+			ViewData["DateSortParm"] = sortOrder == "Date" ? "date_desc" : "Date";
+
+			if (searchString != null)
+			{
+				page = 1;
+			}
+			else
+			{
+				searchString = currentFilter;
+			}
+
+			ViewData["CurrentFilter"] = searchString;
+
+			var students = from s in _context.Students
+						   select s;
+
+			if (!string.IsNullOrEmpty(searchString))
+			{
+				students = students.Where(s => s.LastName.Contains(searchString)
+				|| s.FirstMidName.Contains(searchString));
+			}
+
+			switch (sortOrder)
+			{
+				case "name_desc":
+					students = students.OrderByDescending(s => s.LastName);
+					break;
+				case "Date":
+					students = students.OrderBy(s => s.EnrollmentDate);
+					break;
+				case "date_desc":
+					students = students.OrderByDescending(s => s.EnrollmentDate);
+					break;
+				default:
+					students = students.OrderBy(s => s.LastName);
+					break;
+			}
+			int pageSize = 3;
+			return View(await PaginatedList<Student>.CreateAsync(students.AsNoTracking(), page ?? 1, pageSize));
         }
 
         // GET: Students/Details/5
@@ -57,7 +101,8 @@ namespace ContosoUniversity.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("LastName,FirstMidName,EnrollmentDate")] Student student)
+        public async Task<IActionResult> Create(
+			[Bind("LastName,FirstMidName,EnrollmentDate")] Student student)
         {
 			try
 			{
@@ -132,7 +177,7 @@ namespace ContosoUniversity.Controllers
 		}
 
         // GET: Students/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -140,11 +185,19 @@ namespace ContosoUniversity.Controllers
             }
 
             var student = await _context.Students
+				.AsNoTracking()
                 .SingleOrDefaultAsync(m => m.ID == id);
             if (student == null)
             {
                 return NotFound();
             }
+
+			if (saveChangesError.GetValueOrDefault())
+			{
+				ViewData["ErrorMessage"] =
+					"Delete failed. Try again, and if the problem persists " +
+					"see your system administrator.";
+			}
 
             return View(student);
         }
@@ -154,15 +207,28 @@ namespace ContosoUniversity.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var student = await _context.Students.SingleOrDefaultAsync(m => m.ID == id);
-            _context.Students.Remove(student);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
+            var student = await _context.Students
+				.AsNoTracking()
+				.SingleOrDefaultAsync(m => m.ID == id);
 
-        private bool StudentExists(int id)
-        {
-            return _context.Students.Any(e => e.ID == id);
+			if (student == null)
+			{
+				return RedirectToAction(nameof(Index));
+			}
+
+			try
+			{
+				_context.Students.Remove(student);
+				await _context.SaveChangesAsync();
+				return RedirectToAction(nameof(Index));
+			}
+
+			catch (DbUpdateException /* ex */)
+			{
+				//log the error (uncommnet ex variable name and write a log.)
+				return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+			}
+          
         }
     }
 }
